+++ Ctracker.hpp
@@ -5,10 +5,10 @@
 class CTrack
 {
 public:
-    vector<cv::Point2d> trace;
+    vector<cv::Point2f> trace;
     static size_t NextTrackID;
     size_t skipped_frames;
-    cv::Point2d prediction;
+    cv::Point2f prediction;
     TKalmanFilter *KF;
     CTrack( cv::Point2f p, float dt, float Accel_noise_mag );
     ~CTrack();
@@ -24,7 +24,7 @@ public:
     int maximum_allowed_skipped_frames;
     int max_trace_length;
     vector<CTrack *> tracks;
-    void Update( vector<cv::Point2d> &detections );
+    void Update( vector<cv::Point2f> &detections );
     CTracker( float _dt, float _Accel_noise_mag, double _dist_thres = 60, int _maximum_allowed_skipped_frames = 10, int _max_trace_length = 10 );
     ~CTracker( void );
 };
